Fixed button event listener bug and changed title
Previously when holding down left click on a button and moving outside of the button container, the highlight color would stay until you went back into the button and clicked again. This was resolved using mousemove, and now everytime you click a button it will check if you're still in the container.