Refactor operator handling to support continuous operations
If another operator is pushed before pressing equal, the second to last function pushed to the array will be used to calculate the stotal of num1 and num2. If equal is pressed the last fune equals, the second to last element of the operatorArr is called to evaluate the total. If equal is pushed after two numbers and an operator are entered  the last element of operatorArr is used.