Add click listeners on operator buttons to highlight when chosen and add delay feature
When clicking on operators the border will thicken to show that it's selected and will unselect when hitting clear or enter. Also, a delay feature is added modeling the apple calculator which provides a better user experience when buttons operators are hit or when results are displayed. For example, if user entered 81 / 9 and got 9, the delay feature shows the user that the operation completed. If there is no delay, the 9 would have stayed the same and the user might question if the solution was displayed.