Fix continuous multiplication/division bug and add click effects to all keyboard inputs
Previously, when dividing or multiplying and continuing with a solution with another operator the solution would just go to zero. Now, the calculator will wait to operate with the solution until the next operator is pushed. Also, click effects similar to pushing the keys is added providing a better user-experience. Operators are also selected with thickened borders when pushed through keyboard inputs as well.